pyproject.toml
[project]
name = "ddd_template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.116.2",
    "sqlalchemy[asyncio]==2.0.36",
    "aiosqlite==0.19.0",
    "passlib==1.7.4",
    "pydantic-settings==2.11.0",
    "pyjwt==2.10.1",
    "uvicorn[standard]==0.35.0",
    "minio==7.2.16",
    "python-multipart==0.0.20",
    "bcrypt==3.2.2",
]

[tool.poetry.group.test.dependencies]
pytest = "8.4.2"
pytest-asyncio = "1.2.0"
pytest-mock = "3.15.1"
pytest-cov = "6.3.0"
httpx = "0.28.1"
fastapi = {extras = ["testing"], version = "0.116.2"}
flake8 = "7.3.0"


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
